backend/benigma/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.6
django-filter==25.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Markdown==3.7
psycopg==3.2.5
psycopg-binary==3.2.5
pycparser==2.22
PyJWT==2.9.0
python-gnupg==0.5.4
ruff==0.9.9
sqlparse==0.5.3
tomli==2.2.1
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
